+++ contracts/src/main/kotlin/com/r3corda/contracts/CommercialPaper.kt
@@ -117,7 +117,7 @@ class CommercialPaper : Contract {
                         // Don't allow people to issue commercial paper under other entities identities.
                         "the issuance is signed by the claimed issuer of the paper" by
                                 (output.issuance.party.owningKey in command.signers)
-                        "the face value is not zero" by (output.faceValue.pennies > 0)
+                        "the face value is not zero" by (output.faceValue.quantity > 0)
                         "the maturity date is not in the past" by (time < output.maturityDate)
                         // Don't allow an existing CP state to be replaced by this issuance.
                         // TODO: Consider how to handle the case of mistaken issuances, or other need to patch.