+++ node/src/main/kotlin/com/r3corda/node/internal/AbstractNode.kt
@@ -20,9 +20,6 @@ import com.r3corda.node.services.api.AcceptsFileUpload
 import com.r3corda.node.services.api.CheckpointStorage
 import com.r3corda.node.services.api.MonitoringService
 import com.r3corda.node.services.api.ServiceHubInternal
-import com.r3corda.node.services.transactions.InMemoryUniquenessProvider
-import com.r3corda.node.services.transactions.NotaryService
-import com.r3corda.node.services.transactions.TimestampChecker
 import com.r3corda.node.services.clientapi.NodeInterestRates
 import com.r3corda.node.services.config.NodeConfiguration
 import com.r3corda.node.services.identity.InMemoryIdentityService
@@ -36,6 +33,10 @@ import com.r3corda.node.services.persistence.NodeAttachmentService
 import com.r3corda.node.services.persistence.PerFileCheckpointStorage
 import com.r3corda.node.services.persistence.StorageServiceImpl
 import com.r3corda.node.services.statemachine.StateMachineManager
+import com.r3corda.node.services.transactions.InMemoryUniquenessProvider
+import com.r3corda.node.services.transactions.NotaryService
+import com.r3corda.node.services.transactions.SimpleNotaryService
+import com.r3corda.node.services.transactions.ValidatingNotaryService
 import com.r3corda.node.services.wallet.NodeWalletService
 import com.r3corda.node.utilities.AddOrRemove
 import com.r3corda.node.utilities.AffinityExecutor
@@ -87,7 +88,7 @@ abstract class AbstractNode(val dir: Path, val configuration: NodeConfiguration,
         override val keyManagementService: KeyManagementService get() = keyManagement
         override val identityService: IdentityService get() = identity
         override val monitoringService: MonitoringService = MonitoringService(MetricRegistry())
-        override val clock: Clock get() = platformClock
+        override val clock: Clock = platformClock
     }
 
     val info: NodeInfo by lazy {
@@ -106,6 +107,8 @@ abstract class AbstractNode(val dir: Path, val configuration: NodeConfiguration,
     lateinit var identity: IdentityService
     lateinit var net: MessagingService
     lateinit var api: APIServer
+    var isPreviousCheckpointsPresent = false
+        private set
 
     /** Completes once the node has successfully registered with the network map service. Null until [start] returns. */
     @Volatile var networkMapRegistrationFuture: ListenableFuture<Unit>? = null
@@ -123,27 +126,35 @@ abstract class AbstractNode(val dir: Path, val configuration: NodeConfiguration,
         storage = storageServices.first
         checkpointStorage = storageServices.second
         net = makeMessagingService()
-        smm = StateMachineManager(services, checkpointStorage, serverThread)
         wallet = NodeWalletService(services)
         keyManagement = E2ETestKeyManagementService()
         makeInterestRatesOracleService()
-        api = APIServerImpl(this)
-
-        // Build services we're advertising
-        if (NetworkMapService.Type in info.advertisedServices) makeNetworkMapService()
-        if (NotaryService.Type in info.advertisedServices) makeNotaryService()
-
         identity = makeIdentityService()
+        api = APIServerImpl(this)
+        smm = StateMachineManager(services, listOf(storage, net, wallet, keyManagement, identity, platformClock), checkpointStorage, serverThread)
 
         // This object doesn't need to be referenced from this class because it registers handlers on the network
         // service and so that keeps it from being collected.
         DataVendingService(net, storage)
 
+        buildAdvertisedServices()
+
         startMessagingService()
         networkMapRegistrationFuture = registerWithNetworkMap()
+        isPreviousCheckpointsPresent = checkpointStorage.checkpoints.any()
+        smm.start()
         started = true
         return this
     }
+
+    private fun buildAdvertisedServices() {
+        val serviceTypes = info.advertisedServices
+        if (NetworkMapService.Type in serviceTypes) makeNetworkMapService()
+
+        val notaryServiceType = serviceTypes.singleOrNull { it.isSubTypeOf(NotaryService.Type) }
+        if (notaryServiceType != null) makeNotaryService(notaryServiceType)
+    }
+
     /**
      * Register this node with the network map cache, and load network map from a remote service (and register for
      * updates) if one has been supplied.
@@ -197,10 +208,15 @@ abstract class AbstractNode(val dir: Path, val configuration: NodeConfiguration,
         inNodeNetworkMapService = InMemoryNetworkMapService(net, reg, services.networkMapCache)
     }
 
-    open protected fun makeNotaryService() {
+    open protected fun makeNotaryService(type: ServiceType) {
         val uniquenessProvider = InMemoryUniquenessProvider()
         val timestampChecker = TimestampChecker(platformClock, 30.seconds)
-        inNodeNotaryService = NotaryService(net, storage.myLegalIdentity, storage.myLegalIdentityKey, uniquenessProvider, timestampChecker)
+
+        inNodeNotaryService = when (type) {
+            is SimpleNotaryService.Type -> SimpleNotaryService(smm, net, timestampChecker, uniquenessProvider)
+            is ValidatingNotaryService.Type -> ValidatingNotaryService(smm, net, timestampChecker, uniquenessProvider)
+            else -> null
+        }
     }
 
     lateinit var interestRatesService: NodeInterestRates.Service
@@ -243,7 +259,7 @@ abstract class AbstractNode(val dir: Path, val configuration: NodeConfiguration,
         val checkpointStorage = PerFileCheckpointStorage(dir.resolve("checkpoints"))
         _servicesThatAcceptUploads += attachments
         val (identity, keypair) = obtainKeyPair(dir)
-        return Pair(constructStorageService(attachments, keypair, identity),checkpointStorage)
+        return Pair(constructStorageService(attachments, keypair, identity), checkpointStorage)
     }
 
     protected open fun constructStorageService(attachments: NodeAttachmentService, keypair: KeyPair, identity: Party) =