+++ node/src/main/kotlin/com/r3corda/node/services/identity/InMemoryIdentityService.kt
@@ -2,6 +2,7 @@ package com.r3corda.node.services.identity
 
 import com.r3corda.core.crypto.Party
 import com.r3corda.core.node.services.IdentityService
+import com.r3corda.core.serialization.SingletonSerializeAsToken
 import java.security.PublicKey
 import java.util.concurrent.ConcurrentHashMap
 import javax.annotation.concurrent.ThreadSafe
@@ -10,7 +11,7 @@ import javax.annotation.concurrent.ThreadSafe
  * Simple identity service which caches parties and provides functionality for efficient lookup.
  */
 @ThreadSafe
-class InMemoryIdentityService() : IdentityService {
+class InMemoryIdentityService() : SingletonSerializeAsToken(), IdentityService {
     private val keyToParties = ConcurrentHashMap<PublicKey, Party>()
     private val nameToParties = ConcurrentHashMap<String, Party>()