+++ node/src/main/kotlin/com/r3corda/node/services/messaging/ArtemisMessagingService.kt
@@ -4,8 +4,9 @@ import com.google.common.net.HostAndPort
 import com.r3corda.core.RunOnCallerThread
 import com.r3corda.core.ThreadBox
 import com.r3corda.core.messaging.*
-import com.r3corda.node.internal.Node
+import com.r3corda.core.serialization.SingletonSerializeAsToken
 import com.r3corda.core.utilities.loggerFor
+import com.r3corda.node.internal.Node
 import org.apache.activemq.artemis.api.core.SimpleString
 import org.apache.activemq.artemis.api.core.TransportConfiguration
 import org.apache.activemq.artemis.api.core.client.*
@@ -52,7 +53,7 @@ import javax.annotation.concurrent.ThreadSafe
  */
 @ThreadSafe
 class ArtemisMessagingService(val directory: Path, val myHostPort: HostAndPort,
-                              val defaultExecutor: Executor = RunOnCallerThread) : MessagingService {
+                              val defaultExecutor: Executor = RunOnCallerThread) : SingletonSerializeAsToken(), MessagingService {
     // In future: can contain onion routing info, etc.
     private data class Address(val hostAndPort: HostAndPort) : SingleMessageRecipient
 
@@ -124,9 +125,8 @@ class ArtemisMessagingService(val directory: Path, val myHostPort: HostAndPort,
         val secManager = ActiveMQJAASSecurityManager(InVMLoginModule::class.java.name, secConfig)
         mq.setSecurityManager(secManager)
 
-        // Currently we cannot find out if something goes wrong during startup :( This is bug ARTEMIS-388 filed by me.
+        // TODO Currently we cannot find out if something goes wrong during startup :( This is bug ARTEMIS-388 filed by me.
         // The fix should be in the 1.3.0 release:
-        //
         // https://issues.apache.org/jira/browse/ARTEMIS-388
         mq.start()
 
@@ -136,12 +136,13 @@ class ArtemisMessagingService(val directory: Path, val myHostPort: HostAndPort,
 
         // Create a queue on which to receive messages and set up the handler.
         session = clientFactory.createSession()
+
         session.createQueue(myHostPort.toString(), "inbound", false)
         inboundConsumer = session.createConsumer("inbound").setMessageHandler { message: ClientMessage ->
             // This code runs for every inbound message.
             try {
                 if (!message.containsProperty(TOPIC_PROPERTY)) {
-                    log.warn("Received message without a ${TOPIC_PROPERTY} property, ignoring")
+                    log.warn("Received message without a $TOPIC_PROPERTY property, ignoring")
                     return@setMessageHandler
                 }
                 val topic = message.getStringProperty(TOPIC_PROPERTY)
@@ -159,6 +160,8 @@ class ArtemisMessagingService(val directory: Path, val myHostPort: HostAndPort,
 
                 deliverMessage(msg)
             } finally {
+                // TODO the message is delivered onto an executor and so we may be acking the message before we've
+                // finished processing it
                 message.acknowledge()
             }
         }