+++ docs/source/release-notes.rst
@@ -10,6 +10,8 @@ Here are changes in git master that haven't yet made it to a snapshot release:
 
 * The cash contract has moved from com.r3corda.contracts to com.r3corda.contracts.cash.
 * Amount class is now generic, to support non-currency types (such as assets, or currency with additional information).
+* Refactored the Cash contract to have a new FungibleAsset superclass, to model all countable assets that can be merged
+  and split (currency, barrels of oil, etc.)
 
 
 Milestone 0